Fixed graphiteServer rest calls
