+++ grails-app/conf/application.yml
@@ -318,6 +318,7 @@ grails:
                             {pattern: '/landing/**',                                    access: ["permitAll"]},
                             {pattern: '/applicationDashboard/rest/saveJobHealthGraphiteServers',  access: ['ROLE_ADMIN', 'ROLE_SUPER_ADMIN']},
                             {pattern: '/applicationDashboard/rest/removeJobHealthGraphiteServers',  access: ['ROLE_ADMIN', 'ROLE_SUPER_ADMIN']},
+                            {pattern: '/applicationDashboard/rest/createGraphiteServer',  access: ['ROLE_ADMIN', 'ROLE_SUPER_ADMIN']},
                             {pattern: '/applicationDashboard/rest/createCsiConfiguration', access: ['ROLE_ADMIN', 'ROLE_SUPER_ADMIN']},
                             {pattern: '/applicationDashboard/**',                       access: ["permitAll"]},
                             {pattern: '/queueDashboard/**',                             access: ["permitAll"]},
