+++ grails-app/controllers/de/iteratec/osm/UrlMappings.groovy
@@ -34,7 +34,6 @@ class UrlMappings {
         "/applicationDashboard/rest/$action"(controller: "applicationDashboard")
         "/queueDashboard/rest/$action"(controller: "queueDashboard")
         "/queueDashboard"(view: "/angularFrontend")
-        "/graphiteServer/rest/$action"(controller: "graphiteServer")
 
         ////////////////////////////////////////////////////////////////////////////////////////////////
         // Pages with controller
