+++ grails-app/controllers/de/iteratec/osm/result/ApplicationDashboardController.groovy
@@ -1,10 +1,13 @@
 package de.iteratec.osm.result
 
+import de.iteratec.osm.annotations.RestAction
 import de.iteratec.osm.api.dto.ApplicationCsiDto
 import de.iteratec.osm.api.dto.PageCsiDto
+import de.iteratec.osm.measurement.environment.wptserver.Protocol
 import de.iteratec.osm.measurement.schedule.JobGroup
 import de.iteratec.osm.measurement.schedule.JobGroupService
 import de.iteratec.osm.report.external.GraphiteServer
+import de.iteratec.osm.report.external.provider.GraphiteSocketProvider
 import de.iteratec.osm.util.ControllerUtils
 import grails.validation.Validateable
 import grails.converters.JSON
@@ -100,6 +103,24 @@ class ApplicationDashboardController {
         return ControllerUtils.sendObjectAsJSON(response, message)
     }
 
+    def createGraphiteServer() {
+        try {
+            GraphiteServer server = new GraphiteServer(
+                    serverAdress: params.address,
+                    port: Integer.parseInt(params.port),
+                    webappPathToRenderingEngine: "render",
+                    webappProtocol: Protocol.HTTPS,
+                    reportProtocol: GraphiteSocketProvider.Protocol.valueOf(params.protocol),
+                    webappUrl: params.webAppAddress,
+                    prefix: params.prefix
+            )
+            server.save(failOnError: true, flush: true)
+            return ControllerUtils.sendObjectAsJSON(response, [success: true, id: server.id])
+        }
+        catch(e) {
+            return ControllerUtils.sendObjectAsJSON(response, [success: false, id: null])
+        }
+    }
 
     /**
      * Rest controller that returns all jobs where at least one of the last five measurements failed.
